# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Micro Credits Backend" ,
    "description": "Micro Credits Backend" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "ec2-54-171-157-12.eu-west-1.compute.amazonaws.com:3005",
...
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths":{
...
    "/country/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of country that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete country with given ID",
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "Country is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update country with give ID",
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "country",
            "in": "body",
            "description": "Country with new values of properties",
            "schema": {
              "$ref": "#/definitions/countryUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Country is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Country"
        ],
        "summary": "Get country by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/state": {
      "get": {
        "tags": [
          "State"
        ] ,
        "summary": "Get all states" ,
        "responses": {
          "200": {
            "description": "OK" ,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "body",
            "description": "Body of state that we want to create",
            "schema": {
              "$ref": "#/definitions/stateData"
            }
          }
        ],
        "summary": "Creates new State",
        "responses": {
          "200": {
            "description": "New state is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/state/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of state that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete state with given ID",
        "tags": [
          "State"
        ],
        "responses": {
          "200": {
            "description": "State is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update state with give ID",
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "body",
            "description": "State with new values of properties",
            "schema": {
              "$ref": "#/definitions/stateUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "State"
        ],
        "summary": "Get state by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/state/country/{countryID}": {
      "parameters": [
        {
          "name": "countryID",
          "in": "path",
          "required": true,
          "description": "ID of country that we want to find states of",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "State"
        ],
        "summary": "Get states by country id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
    "/city/state/{stateID}": {
      "parameters": [
        {
          "name": "stateID",
          "in": "path",
          "required": true,
          "description": "ID of state that we want to find cities of",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Get cities by state id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "countryUpdate": {
      "required": [
        "countryName"
      ],
      "properties": {
        "countryName": {
          "type": "string"
        }
      }
    },
    "stateData": {
      "required": [
        "stateName",
        "countryId"
      ],
      "properties": {
        "stateName": {
          "type": "string"
        },
        "countryId": {
          "type": "integer"
        }
      }
    },
    "stateUpdate": {
      "required": [
        "stateName"
      ],
      "properties": {
        "stateName": {
          "type": "string"
        }
      }
    },
...
  }
}